Breakdown tsdb/head.go into multiple files
